test: address PR #214 reviewer follow-ups (batch 2)
Bug fixes:

- GLStateGuard sizes bumped 16->32 for both texture slots and UBO bindings; engine already uses unit 24 (shadow/IBL/foliage) and UBO 18 (post-process fog), so the previous limits missed real leaks. Capture loops now also clamp against GL_MAX_TEXTURE_IMAGE_UNITS / GL_MAX_UNIFORM_BUFFER_BINDINGS so software renderers with fewer slots do not over-query.

- RendererValidate::IsFloatColorFormat tightened to RGBA-only. Previously accepted RG/RGB float formats but the readback always issued glGetTextureImage(GL_RGBA), which is invalid for non-RGBA base formats.

- RenderGraph::UpdateDependencyGraph now returns bool; ValidateResourceHazards and Execute abort on cycles instead of continuing with a partial m_PassOrder (the prior behaviour produced misleading 'missing dependency' reports for passes the cycle had excluded from the topo sort).

- MeshBinarySerializer bbox fix still intact from previous commit (mentioned for completeness).

Test hygiene:

- FuzzAnimationBinary caps on-disk write at 4 MB (parity with FuzzAssimpMesh).

- FuzzAssimpMesh comment updated to reflect per-hint importer recreation.

- FuzzInputActionYaml / FuzzMeshBinary / FuzzAnimationBinary / FunctionWithContextTest: explicit <string> / <system_error> includes.

- ResourceHandle.h: explicit <functional> include for std::hash.

- ShaderUnitTests::RunToneMap now ADD_FAILURE + early-return on shader compilation failure instead of EXPECT_TRUE then dereferencing null.

- GoldenImageTests: ASSERT_TRUE on Shader::Create return value before Bind().

- PbrPropertyTests: std::memcpy into aligned f32 buffer instead of reinterpret_cast on readback bytes (aliasing UB).

- DataRoundTripTests: PID-suffixed temp IBL cache dir so parallel runners don't collide.

- MathTest: u32 loop indices and f32 eps per engine typedefs.

- TestFailureCaptureTest: dropped redundant static inside anonymous namespace.

- perf_trend.py: dropped redundant int() around round() (Python 3 round returns int).

Tooling / CI:

- generate_test_catalogue.py: per-layer heading depth '###' (was '####') so it slots correctly under '## 3.12' and satisfies MD001.

- gtest_discover_tests now passes WORKING_DIRECTORY=<repo>/OloEditor so ctest-discovered tests resolve assets/shaders/* paths correctly.

- .gitignore: OloEditor/assets/tests/captures/ is runtime test-failure output; removed six committed machine-specific files from tracking.

- asan.yml: ASan/UBSan/TSan jobs pinned to ubuntu-24.04 (matches the hardcoded LLVM apt repo codename).

- cross-vendor.yml: added inline comment documenting Mesa version + last-verified date.

- fuzz.yml: missing corpus directory now fails the workflow instead of silently running the harness without seeds.

Skipped reviewer items (with rationale):

- lsan.txt broad suppressions: third-party intentional leaks live in many allocation paths; tightening to specific init symbols would unmask benign leaks from paths we don't control. Current module-level suppressions are the idiomatic LSAN practice.

- RenderPropertyTest ChangeToOloEditorDir restore: CWD change is intentional one-time init; restoring breaks every subsequent test that resolves assets via relative path.

- PbrPropertyTests / PostProcessPropertyTests GLStateGuard wrapping: tests intentionally mutate GL state; the guard is for production passes, not test harnesses.

- EnsureGpuOrSkip macro->method: GTEST_SKIP must run in the test body; the macro is the idiomatic gtest wrapper.

- TestFailureCapture SafeGlString / HasGlContext / WriteCurrentDrawFboPng rework: runs only on test failure; current best-effort error-clear + viewport readback works for every real failure we've captured.

- Texture helpers GLuint migration: pure stylistic churn, u32 is already the engine typedef and GLuint is interchangeable.

+++ OloEngine/tests/scripts/perf_trend.py
@@ -37,7 +37,7 @@ def percentile(values, q):
     if not values:
         return float("nan")
     s = sorted(values)
-    k = max(0, min(len(s) - 1, int(round((q / 100.0) * (len(s) - 1)))))
+    k = max(0, min(len(s) - 1, round((q / 100.0) * (len(s) - 1))))
     return s[k]
